WIP - Protocol Buffer Reading from C# Generated PB File
Still a work in progress. I am running into an issue when reading the binary data on the java side. It will read and parse into PBTinkarMsg objects for only about the first 130 ish. I've referenced/mirrored the working C# code and it's almost identical and just as minimal in implementation.

Co-Authored-By: Kurt W. Allen <5391535+Gjoll@users.noreply.github.com>